Improves Automaton Intrusion messages
Updates the Automaton Intrusion messages to include the name of the
user who didn't secure their machine properly and caused the breach.

This provides more context to the players and encourages them to
secure their machines.

Also, the available channels are now different depending on the bot
(dev or prod).